1d profile cutting optimization
Hello Gentlemen,
I wanted to make a VBA that creates self-renewing combinations and I actually did it, but Excel was crashing when the combinations increased, unfortunately Excel does not work in multi-item lists. That's why I worked out the codes that create the combinations with C#.
I am adding all my work, I think someone who understands software should develop this further.
current operation.
Cutting lengths in Excel are collected in a list.
C# creates self-renewing combinations from numbers whose sum reaches the stock size.
selects the largest one and prints it to Excel.
The numbers entered in Excel are deleted from the first list and the process is repeated for the remaining cutting dyes. The process repeats until there is no cutting size left in the first list.
Remember, I am not a software developer, these codes need to be improved.


Please test, we can do better together.
Please review the shortcomings and errors of the codes and let me know the mistakes and shortcomings and teach me the correct one.

Result: It meets my needs.